Fix dashboard test and refactor test authentication
This commit introduces a few changes to mirage setup. First thing is
that instead of hardcode the user in `/users/me` endpoint I fetch the
user based on the `Authorization` header content. This in turn can be
set when using `authenticateSession` helper from Ember Simple Auth.

Another thing is that in order to return proper JSON API formatted
response we need to either let Mirage serialize the response or
serialize it manually. It seems that returning a single record (like
`return user` in `/users/me` handler) does the job. It doesn't work for
collection (like `return user.teams`), so I needed to run
`this.serialize` explicitly. It might be worth investigating what's the
problem to not have to explicitly serialize.

Another thing is that I explicitly defined application serializer to be
a subclass of JSON API Serialier from Ember Data.